Move codecheck fixes from YZ to BZ
Description: Move codecheck fixes from YZ to BZ.
Fix implicit casts, unitialized variables, unsafe functions, check return value

Issue: #I8WZMC

Testing: Build

+++ static_core/compiler/optimizer/analysis/monitor_analysis.cpp
@@ -53,7 +53,7 @@ void MonitorAnalysis::MarkedMonitorRec(BasicBlock *bb, int32_t numMonitors)
             --numMonitors;
         }
     }
-    enteredMonitorsCount_->at(bb->GetId()) = numMonitors;
+    enteredMonitorsCount_->at(bb->GetId()) = static_cast<uint32_t>(numMonitors);
     if (numMonitors == 0) {
         return;
     }
